Added a new source from Frank McCosker
Report by WEF on Internet for all produced in 2016 and Paul Garnet and Paul Mitchell were contributors. 
This document, part of the Access/Adoption pillar within the World Economic Forum’s Global Challenge initiative on the Future of the Internet, provides a common framework for addressing the barriers related to achieving an internet for all. The document is also a foundation for their focus in 2016 and beyond to develop new scalable and replicable on-the-ground collaborative models, in cooperation with governments, that accelerate the achievement of the broader social and economic priorities of a country or region in the context of accelerating internet access and adoption.